refactor(components): subdivide the files
