+++ src/components/Location/LocationResult/index.js
@@ -1,0 +1,3 @@
+import LocationResult from './LocationResult';
+
+export default LocationResult;
