+++ src/components/Location/LocationSearch/LocationSearch.jsx
@@ -1,10 +1,10 @@
-import { memo } from 'react';
-import isEqual from 'react-fast-compare';
 import locationEditAtom from '@Recoil/location/locationEdit';
 import locationValueAtom from '@Recoil/location/locationValue';
+import { memo } from 'react';
+import isEqual from 'react-fast-compare';
 import { useRecoilState } from 'recoil';
 
-import * as S from '../Location.style';
+import * as S from './LocationSearch.style';
 
 const LocationSearch = () => {
   const [value, setValue] = useRecoilState(locationValueAtom);
