+++ src/components/Location/LocationSearch/index.js
@@ -1,0 +1,3 @@
+import LocationSearch from './LocationSearch';
+
+export default LocationSearch;
